+++ wayland/unstable/keyboard-shortcuts-inhibit-v1/keyboard_shortcuts_inhibit.go
@@ -262,7 +262,7 @@ func (i *KeyboardShortcutsInhibitor) AddInactiveHandler(f KeyboardShortcutsInhib
 	i.inactiveHandlers = append(i.inactiveHandlers, f)
 }
 
-func (i *KeyboardShortcutsInhibitor) Dispatch(opcode uint16, fd uintptr, data []byte) {
+func (i *KeyboardShortcutsInhibitor) Dispatch(opcode uint32, fd int, data []byte) {
 	switch opcode {
 	case 0:
 		if len(i.activeHandlers) == 0 {
